libs/checkpoint-kusto/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langgraph-checkpoint-kusto"
version = "1.0.0"
description = "Azure Data Explorer (Kusto) implementation of LangGraph checkpoint saver."
authors = []
requires-python = ">=3.10"
readme = "README.md"
license = "MIT"
license-files = ['LICENSE']
dependencies = [
  "langgraph-checkpoint>=2.1.2,<4.0.0",
  "orjson>=3.10.1",
  "azure-kusto-data[aio]>=4.3.1",
  "azure-kusto-ingest>=4.3.1",
  "azure-identity>=1.15.0",
  "aiohttp>=3.8.0",
]

[project.urls]
Repository = "https://www.github.com/langchain-ai/langgraph"

[dependency-groups]
test = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-mock>=3.11.0",
  "anyio",
  "langgraph-checkpoint",
  "pytest-watcher",
]
lint = [
  "ruff>=0.1.0",
  "codespell>=2.2.0",
  "mypy>=1.7.0",
]
dev = [
  {include-group = "test"},
  {include-group = "lint"},
]

[tool.uv]
default-groups = ['dev']

[tool.uv.sources]
langgraph-checkpoint = { path = "../checkpoint", editable = true }

[tool.hatch.build.targets.wheel]
include = ["langgraph"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --durations=5 -vv"
asyncio_mode = "auto"

[tool.ruff]
lint.select = [
  "E",  # pycodestyle
  "F",  # Pyflakes
  "UP", # pyupgrade
  "B",  # flake8-bugbear
  "I",  # isort
]
lint.ignore = ["E501", "B008"]
target-version = "py310"

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html
disallow_untyped_defs = "True"
explicit_package_bases = "True"
warn_no_return = "False"
warn_unused_ignores = "True"
warn_redundant_casts = "True"
allow_redefinition = "True"
disable_error_code = "typeddict-item, return-value"

[tool.pytest-watcher]
now = true
delay = 0.1
runner_args = ["--ff", "-x", "-v", "--tb", "short"]
patterns = ["*.py"]
